fix 인덱스어레이의 자료형과 인덱스버퍼뷰의 포멧이 일치하지않아서 모델이 뜨지 않는 오류 해결 feat 모델 띄우기 성공

+++ GOE/GOE_Render/Kuramon.h
@@ -55,10 +55,9 @@ public:
 	/// 
 	/// </summary>
 	std::vector<Vertex> vertexArray;
-	std::vector<UINT16> indexArray;
+	std::vector<UINT32> indexArray;
 	MeshData* m_kuramonMeshData = nullptr;
 
-	//Vertex m_kuramonTriangleVertices[6180] = {};
 	std::vector<Vertex> m_kuramonTriangleVertices; // 얘만 어떻게 하면 될지도? 
 	UINT m_vertexBufferSize = 0;
 	ComPtr<ID3D12Resource> m_kuramonVertexBufferUpload = nullptr;
